Extract WebGPU mipmap and clear renderer shaders into shader chunks
