+++ src/platform/graphics/webgpu/webgpu-mipmap-renderer.js
@@ -2,6 +2,7 @@ import { Shader } from '../shader.js';
 import { SHADERLANGUAGE_WGSL } from '../constants.js';
 import { Debug, DebugHelper } from '../../../core/debug.js';
 import { DebugGraphics } from '../debug-graphics.js';
+import webgpuMipmap from '../shader-chunks/frag/webgpu-mipmap.js';
 
 /**
  * @import { WebgpuGraphicsDevice } from './webgpu-graphics-device.js'
@@ -29,41 +30,12 @@ class WebgpuMipmapRenderer {
     constructor(device) {
         this.device = device;
 
-        // Shader that renders a fullscreen textured quad
-        const code = `
- 
-            var<private> pos : array<vec2f, 4> = array<vec2f, 4>(
-                vec2(-1.0, 1.0), vec2(1.0, 1.0),
-                vec2(-1.0, -1.0), vec2(1.0, -1.0)
-            );
-
-            struct VertexOutput {
-                @builtin(position) position : vec4f,
-                @location(0) texCoord : vec2f
-            };
-
-            @vertex
-            fn vertexMain(@builtin(vertex_index) vertexIndex : u32) -> VertexOutput {
-              var output : VertexOutput;
-              output.texCoord = pos[vertexIndex] * vec2f(0.5, -0.5) + vec2f(0.5);
-              output.position = vec4f(pos[vertexIndex], 0, 1);
-              return output;
-            }
-
-            @group(0) @binding(0) var imgSampler : sampler;
-            @group(0) @binding(1) var img : texture_2d<f32>;
-
-            @fragment
-            fn fragmentMain(@location(0) texCoord : vec2f) -> @location(0) vec4f {
-              return textureSample(img, imgSampler, texCoord);
-            }
-        `;
-
+        // shader that renders a fullscreen textured quad
         this.shader = new Shader(device, {
             name: 'WebGPUMipmapRendererShader',
             shaderLanguage: SHADERLANGUAGE_WGSL,
-            vshader: code,
-            fshader: code
+            vshader: webgpuMipmap,
+            fshader: webgpuMipmap
         });
 
         // using minified rendering, so that's the only filter mode we need to set.
